requirements.txt
dependency_injector
Flask
flask_cors
flask_socketio
flask_sqlalchemy
fuzzywuzzy
jieba
numpy
pandas
pydub
python-dotenv
torch
vertexai
google-cloud-speech
webrtcvad-wheels
Werkzeug
python-Levenshtein
mecab-python3
unidic-lite
psycopg2-binary
noisereduce
librosa
git+https://github.com/openai/whisper.git@517a43